♻️REFACTOR: Refactors event drawing from interview manager to dedicated class
