+++ BaysianBeliefNetworkDemo/Assets/Scripts/Interviews/EventDrawer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a4adca78cbf05a9459326c506213fadc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
